Data were normalized incorrectly
Mean subtraction (a.k.a. "mean centering") is necessary for performing
classical PCA to ensure that the first principal component describes the
direction of maximum variance If mean subtraction is not performed, the
first principal component might instead correspond more or less to the
mean of the data. A mean of zero is needed for finding a basis that
minimizes the mean square error of the approximation of the data.
-- from wikipedia

Drop normalization in LBP.
Scale LBP data to similar magnitude of image.

Subtract mean of train data from test data before classification

Also play with the eigenfaces

Signed-off-by: Guobin Zhang <guobin.zhang@intel.com>